Break up app.rs into a top_panel.rs, rerun_menu.rs and, loading.rs
